Serialization fixes after switch to SAX-style JSON writing API
1. Continue to serialize the `cputype` / `cpusubtype` fields of
   `mach_header_64` as unsigned, despite them being signed. This
   preserves compatibility with the existing metadata version.
2. Add the `Serialize` declaration for the special `std::pair<uint64_t,
   std::pair<uint64_t, uint64_t>>` overload to the header. This ensures
   it will be favored over the generic `std::pair<First, Second>`
   template function and preserves the serialization used with the
   existing metadata version.

+++ view/sharedcache/core/MetadataSerializable.cpp
@@ -206,8 +206,11 @@ void Deserialize(DeserializationContext& context, std::string_view name, std::ve
 void Serialize(SerializationContext& context, const mach_header_64& value) {
 	context.writer.StartArray();
 	Serialize(context, value.magic);
-	Serialize(context, value.cputype);
-	Serialize(context, value.cpusubtype);
+	// cputype and cpusubtype are signed but were serialized as unsigned in
+	// v4.2 (metadata version 2). We continue serializing them as unsigned
+	// so we don't need to bump the metadata version.
+	Serialize(context, static_cast<uint32_t>(value.cputype));
+	Serialize(context, static_cast<uint32_t>(value.cpusubtype));
 	Serialize(context, value.filetype);
 	Serialize(context, value.ncmds);
 	Serialize(context, value.sizeofcmds);
